Fixed Notepad Session (Third time's the charm, eh?)
Also implemented simple rename, cleaned up some bits we weren't going to
use, and renamed the activity log and comments. I'm starting to think
putting the activity log and comments on here may not have been a useful
idea. :p

@Grace : The latest email he sent us said that open, close, write, and
exit do not have to be implemented for the submission on Monday.