Write & Exit done-ish.
This is less like "Integration of the Parser and Engine" and more like
"write one function call in the manager and spend all your time writing
and fixing things we forgot to do earlier". Haha.

Grace : database::write() was an empty function so I implemented it
following your format given in the testfiles such as "age.txt",
"animals.txt", etc. The only bit of confusion I ran into was the line
"Key: <Attribute>" where <Attribute> is an attribute of that table which
I assume to be the primary key in that relation. I wasn't exactly sure
where I would fetch the primary key from so I left you a comment on line
134-137 in Database.cpp.

PS: Write and Exit done-ish.
PPS: Please excuse my inefficient code; I tried my best to aim for
flexibility and readability rather than optimization and efficiency.