Handling queries completed
In order to tie this into your code, call handleQuery(string s) inside
parser with the string that needs to be evaluated. The code will take
care of the rest. The result of the query will be stored in the database
Relation object called temp. (Wes, this is slightly different than what
we talked about.)

I did not finish select but I will do so later in the morning.